You can now password protected Terminals as a whole.  If you do, it will store your password in one-way hash format and you will not be able to use the application without providing that password.  Check the Tools, Options, Security tab to set this password.
It will also use that password as the Key and IV for the Rijndael encryption of the rest of your passwords into the config file.  That way all your passwords  will rely on the single 1 way hash.  That way, you cant simply delete the Terminals Application password to recover the rest.

If you do not provide a password it will use the original DPAPI method.