Enable running from a network location
This configuration setting allows library loading from network locations. If the terminals program is located on a mapped drive or a network, currently it will crash on load due to it not being permitted to access the assemblies in the Plugins folder. This fixes that so that the tool can be run from a network share, as our department does.